Add "context" option to avoid global state
This commit adds a `context: IContext` option to the constructors of
Document and Collection. An object's context contains references
to its Firebase app and Firestore instance, which will be used
instead of the global references in `init.ts`. Collections propagate
context to their child documents.

This allows consumers to use Firestorter with multiple Firebase apps,
which is especially helpful when writing tests using the
'@firebase/testing' module.

No changes are needed to existing code: if `context` is undefined on
a Document or Collection, the global context will be used instead.